Refactor: move isNoSitePresent and ensureSiteIsPresent to multiSite module
