+++ src/multiSite/isNoSitePresent.ts
@@ -1,0 +1,5 @@
+import { getLanguageVersions } from '../config/scrivitoSites'
+
+export function isNoSitePresent(): boolean {
+  return !getLanguageVersions()?.length
+}
